ТУЧА ФЗНАМЗНОН

+++ maxagen.py
@@ -1,14 +1,15 @@
 import discord
-import os
+from distutils.util import strtobool
 import random
 
+
 class MaxxaGen:
-    def __init__(self, settings):
+    def __init__(self, settings, file):
         self.all_phrs = settings["all_phrases"]
         self.line_phrs = settings["line_phrases"]
         self.captlz = settings["capitalize"]
         self.allow_repeats = settings["allow_repeats"]
-        self.file = settings["file"]
+        self.file = file
         self.phrases = self.read_from_file()
 
     def read_from_file(self):
@@ -40,7 +41,7 @@ class MaxxaGen:
         while counter != self.all_phrs:
             counter += 1
 
-            phrase = self.get_random(False) if self.allow_repeats is True else self.get_random(True)
+            phrase = self.get_random(False) if self.allow_repeats else self.get_random(True)
 
             if phrase:
                 if line_counter == self.line_phrs:
@@ -66,26 +67,9 @@ class MaxxaGen:
         for val in res_dict.values():
             if res_str:
                 res_str += "\n"
-            res_str += val.capitalize() if self.captlz is True else val
+            res_str += val.capitalize() if self.captlz else val
 
         return res_str
-
-
-settings_gen = {
-    "all_phrases": 5,
-    "line_phrases": 5,
-    "capitalize": True,
-    "allow_repeats": True,
-    "file": "phrases.txt",
-}
-
-settings_forgiveness = {
-    "all_phrases": 5,
-    "line_phrases": 5,
-    "capitalize": True,
-    "allow_repeats": True,
-    "file": "forgiveness_phrases.txt",
-}
 
 
 client = discord.Client()
@@ -105,22 +89,33 @@ async def on_message(message):
     help_commands = ["help", "помощь", "помогите"]
     hello_commands = ["привет", "здарова", "hello", "йоу"]
 
+    all_cmds = commands + sorry_commands + gen_commands + help_commands + hello_commands
+
     msg_content_splitted = message.content.lower().split(",")
     msg_content = msg_content_splitted[0]
-    #
-    all_count = int(msg_content_splitted[1]) if len(msg_content_splitted) >= 2 else 5
-    line_count = int(msg_content_splitted[2]) if len(msg_content_splitted) >= 3 else 5
-    cap = msg_content_splitted[3] if len(msg_content_splitted) >= 4 else True   # TODO: Converting to bool
-    rep = msg_content_splitted[4].bool if len(msg_content_splitted) == 5 else True
+    # ebatj ja bidlo ebanoe
+    try:
+        all_count = int(msg_content_splitted[1])
+    except Exception:
+        all_count = 5
+    try:
+        line_count = int(msg_content_splitted[2])
+    except Exception:
+        line_count = 5
+    try:
+        cap = strtobool(msg_content_splitted[3])
+    except Exception:
+        cap = True
+    try:
+        rep = strtobool(msg_content_splitted[4])
+    except Exception:
+        rep = False
 
-    # print(all_count, line_count, cap, rep)
-
-    st = {
+    settings = {
         "all_phrases": all_count,
         "line_phrases": line_count,
         "capitalize": cap,
-        "allow_repeats": rep,
-        "file": "phrases.txt",
+        "allow_repeats": rep
     }
 
     # ignore bot messages
@@ -131,41 +126,40 @@ async def on_message(message):
     with open("help.txt", "r", encoding="utf8") as file:
         data = file.read()
 
-
     # mahammad20 messages duplicator
     if message.author.id == 373115287828037632:
         await message.channel.send(f"```{message.content}```")
     # if message.author.id == 373115287828037632:
     #     await message.channel.send(f"```{message.content}```")
-    if msg_content.startswith('/'):
+    if msg_content.startswith('/') and any(ele in msg_content for ele in all_cmds):
         # delete command message
         if msg_content.startswith('/?'):
             await message.delete()
 
         # checks if dolboeb maha writes command without mistakes
-        if message.author.id == 373115287828037632 and message.content.startswith(tuple(commands)) and " " in message.content:
+        if message.author.id == 373115287828037632 and message.content.startswith(
+                tuple(commands)) and " " in message.content:
             await message.channel.send("Махмуд пошёл нахуй, читай команды долбаёб")
         # if maha asks somebody say "jo", bot will print "jo" 60 times in 1 msg
         elif message.author.id == 373115287828037632 and 'пойоукай' in message.content:
-          await message.channel.send('йоу ' * 60)
+            await message.channel.send('йоу ' * 60)
         # $gen command
         elif any(ele in msg_content for ele in gen_commands):
-            print(st)
-            print(MaxxaGen(st))
-            await message.channel.send(MaxxaGen(st))
+            # print(settings_gen)
+            # print(MaxxaGen(settings_gen))
+            await message.channel.send(MaxxaGen(settings, "phrases.txt"))
         # $sry command
         elif any(ele in msg_content for ele in sorry_commands):
-            await message.channel.send(MaxxaGen(settings_forgiveness))
+            await message.channel.send(MaxxaGen(settings, "forgiveness_phrases.txt"))
         # $help command
         elif any(ele in msg_content for ele in help_commands):
             await message.channel.send(f"```{data}```")
         # $hi command
         elif any(ele in msg_content for ele in hello_commands):
-            await message.channel.send("йоу бандишка братишка") # добавить файл с разными приветствиями и принтить рандомное
+            await message.channel.send(MaxxaGen(settings, "hello.txt"))
         # skips written command /play (Rhytm bot) by maha
         elif message.author.id == 373115287828037632 and '/play' in message.content:
             pass
 
 
-
-client.run('ODMxNjQ3OTAyNjkzOTgyMjI4.YHYSdw.i7Eduu37I0bHcHr8bUWjmRRe4lY')
+client.run('ODMxNjQ3OTAyNjkzOTgyMjI4.YHYSdw.QT7y3IEI-tHuXdjeg32e0AqUEGo')
